pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "serial-bin-plotter"
version = "0.1.0"
description = "DiffBot telemetry viewer and serial binary plotter."
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
  "PyQt6",
  "numpy",
  "pyqtgraph",
  "pyserial",
]

[project.scripts]
serial-bin-plotter = "main:main"

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "ui"]
include = [
  "main.py",
  "styles.py",
  "streams.json",
]
